# Project-Travista/docs/api-specification.yaml
openapi: 3.1.0
info:
  title: AI Trip Planner API
  description: Comprehensive travel planning and booking platform API
  version: 1.0.0
  contact:
    name: API Support
    email: api@tripgenius.com
servers:
  - url: https://api.tripgenius.com/v1
    description: Production server
  - url: https://staging-api.tripgenius.com/v1
    description: Staging server
security:
  - BearerAuth: []
paths:
...
  /trips:
    get:
      summary: Get user's trips
      tags: [Trips]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, planned, confirmed, active, completed, cancelled]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of trips
          content:
            application/json:
              schema:
                type: object
                properties:
                  trips:
                    type: array
                    items:
                      $ref: '#/components/schemas/Trip'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      summary: Create new trip
      tags: [Trips]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTripRequest'
      responses:
        '201':
          description: Trip created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
...
  /bookings:
    get:
      summary: Get user's bookings
      tags: [Bookings]
      parameters:
        - name: tripId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, cancelled, completed]
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Booking'

    post:
      summary: Create booking
      tags: [Bookings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Booking initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking:
                    $ref: '#/components/schemas/Booking'
                  paymentUrl:
                    type: string
                    format: uri
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Trip:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        destination:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        partySize:
          type: integer
        budget:
          type: object
          properties:
            min:
              type: number
            max:
              type: number
            currency:
              type: string
        status:
          type: string
          enum: [draft, planned, confirmed, active, completed, cancelled]
        createdAt:
          type: string
          format: date-time
...
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tripId:
          type: string
          format: uuid
        vendor:
          type: string
        type:
          type: string
          enum: [flight, hotel, activity, transport]
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        amount:
          type: number
        currency:
          type: string
        confirmationCode:
          type: string
        documentsUrl:
          type: string
          format: uri
...
    CreateTripRequest:
      type: object
      required: [title, destination, startDate, endDate, partySize]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 100
        destination:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        partySize:
          type: integer
          minimum: 1
        budget:
          type: object
          properties:
            min:
              type: number
            max:
              type: number
            currency:
              type: string
    CreateBookingRequest:
      type: object
      required: [tripId, vendorId, type, items]
      properties:
        tripId:
          type: string
          format: uuid
        vendorId:
          type: string
          format: uuid
        type:
          type: string
          enum: [flight, hotel, activity, transport]
        items:
          type: array
          items:
            type: object
        guestInfo:
          type: object
        paymentMethod:
          type: object
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
...